udev: setup becomes serial_handshake AT 9600 baud
now all arduino-udev enumeration is done at 9600 baud.
this means we can support all devices on long cables
more reliably. the downside is that device startup
can take longer as we wait longer for the blocking
serial_handshake step. python users should use
the corresponding python api for ease of use.

at the arduino end, also default to using the
onboard led to show we are in the handshake phase. this
is visible with rapid blinking. users should call
UDev.begin() and UDev.serial_handshake() as the first
operation in setup()

notable lessons: add a delay(100) before closing the serial
port to ensure the remaining data gets flushed