udev,python: robustness improvements
terribly lazy, but just halve the wait time and
try twice.

after much investigation I think is is an issue of buffering
on the linux side as the way to reproduce startup errors
is to close a the programming after init. after a few
seconds the tx light stops flashing (I presume due to
flow control or other, which I thought was disabled BTW??) the
ftdi stops sending data. the next port open we get a bunch of
rubbish. closing the port and reopening seems to remove this
cruft and the next name request succeeds. it appears the
serial.flushInput() doesn't do enough / the right thing.
even cat /dev/ttyUSB0 before opening the port fixes it.

so to work around all this stuff, just try twice.